+++ genetics/test/test_chromosomes.py
@@ -4,12 +4,19 @@ from typing import Dict
 
 import pytest
 
-from genetics.chromosomes import GenericChromosome
+from genetics.chromosomes import GenericChromosome, GenotypeException
+from genetics.genes import GenericGene
 
 
-def test_empty_chromosome(binary_alphabet: Dict):
-    test_chromosome = GenericChromosome(0, binary_alphabet, genes=[])
-    assert not len(test_chromosome)
+def test_empty_chromosome(binary_alphabet: Dict) -> None:
+    generic_chromosome = GenericChromosome(0, binary_alphabet, genes=[])
+    assert not len(generic_chromosome)
+
+
+def test_size_mismatch_chromosome(binary_alphabet: Dict, size_mismatch_error_msg: str) -> None:
+    with pytest.raises(GenotypeException) as genotype_error:
+        _ = GenericChromosome(0, binary_alphabet, genes=[GenericGene(binary_alphabet)])
+        assert size_mismatch_error_msg in genotype_error.value
 
 
 @pytest.fixture()
@@ -20,6 +27,13 @@ def binary_alphabet() -> Dict[bool, int]:
 @pytest.fixture()
 def ascii_alphabet() -> Dict[int, str]:
     return dict([n for n in enumerate(string.ascii_letters)])
+
+
+@pytest.fixture()
+def size_mismatch_error_msg() -> str:
+    return "Chromosome size doesn't match the number of genes.\n" \
+           "Expected: 0\n" \
+           "Actual: 1"
 
 
 if __name__ == '__main__':