+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 1.0
 
+- (B.9) Test for mismatching size in chromosome
 - (B.8) Test for empty chromosome
 - (B.7) Implemented chromosome class and mutation method
 - (B.6) Test for ASCII gene